Replace Microbits class with new version
Remove unused code.
Remove tests for unused code.
