+++ src/components/connection-prompt/usb/FindUsbDialog.svelte
@@ -7,7 +7,7 @@
 <script lang="ts">
   import StandardButton from '../../StandardButton.svelte';
   import { t } from '../../../i18n';
-  import Microbits from '../../../script/microbit-interfacing/MicrobitsAlt';
+  import Microbits from '../../../script/microbit-interfacing/Microbits';
 
   export let onFoundUsb: () => void;
   export let onUsbLinkError: () => void;
